Make form state changes
